refactor: ghost 관련 타입 api/types/item 으로 이동, 옷장 페이지에서 상태 관리 로직 커스텀 훅으로 분리
