+++ apps/web/app/_api/types/item.ts
@@ -1,6 +1,22 @@
-import { Ghost, Code } from 'app/closet/_fixture/item';
-
 import { ResponseFormat } from './response';
+
+export type Code =
+  | 'mummy_ghost'
+  | 'basic_ghost'
+  | 'cat_ghost'
+  | 'wizard_ghost'
+  | 'detective_ghost'
+  | 'littleprince_ghost'
+  | 'frankenstein_ghost'
+  | 'redhood_ghost'
+  | 'cheese_cat_ghost'
+  | 'siamese_cat_ghost'
+  | 'baekdo_cat_ghost';
+
+export type Ghost = {
+  name: string;
+  code: Code;
+};
 
 export type GetItemListAPIResponse = ResponseFormat<{
   items: {
